Updated Project 2.1 script
I updated the Matlab script. It covers all the required material in the project handout. All theoretical probabilities are calculated. Both the symmetric and non-symmetric channel are implemented with the appropriate probabilities. It also tests both channels for a number of different length bit sequences. I would really appreciate it if you could review the script.

Assuming all is well there, all that needs to be done is to write the report.